+++ CHANGELOG.md
@@ -42,4 +42,7 @@
 ### HDFS writer
 
 * Add support for `timestamp` data type
-* Add support for `lz4`, `zstd` compress format
+
+### MongoDB reader
+
+* Add support for `json` data type